requirements.txt
numpy
absl-py
tensorflow-gpu
scipy
tmuxp
